Load protected packages from installroot
The set of protected packages might differ between systems therefore it
would be good to use a relevant set for the current location.
